+++ service/dataexchange/api.go
@@ -2851,8 +2851,8 @@ func (s *CreateDataSetOutput) SetUpdatedAt(v time.Time) *CreateDataSetOutput {
 
 // The CreateJob request. AWS Data Exchange Jobs are asynchronous import or
 // export operations used to create or copy assets. A data set owner can both
-// import and export as they see fit. Someone with an entitlement to a data
-// set can only export. Jobs are deleted 90 days after they are created. Created
+// import and export assets. A subscriber with an entitlement to a data set
+// can only export. Jobs are deleted 90 days after they are created. Created
 // jobs must be started with the StartJob operation.
 type CreateJobInput struct {
 	_ struct{} `type:"structure"`
@@ -3789,6 +3789,129 @@ func (s *ExportAssetsToS3ResponseDetails) SetEncryption(v *ExportServerSideEncry
 // SetRevisionId sets the RevisionId field's value.
 func (s *ExportAssetsToS3ResponseDetails) SetRevisionId(v string) *ExportAssetsToS3ResponseDetails {
 	s.RevisionId = &v
+	return s
+}
+
+// Details of the operation to be performed by the job.
+type ExportRevisionsToS3RequestDetails struct {
+	_ struct{} `type:"structure"`
+
+	// The unique identifier for the data set associated with this export job.
+	//
+	// DataSetId is a required field
+	DataSetId *string `type:"string" required:"true"`
+
+	// Encryption configuration for the export job.
+	Encryption *ExportServerSideEncryption `type:"structure"`
+
+	// The destination for the revision.
+	//
+	// RevisionDestinations is a required field
+	RevisionDestinations []*RevisionDestinationEntry `type:"list" required:"true"`
+}
+
+// String returns the string representation
+func (s ExportRevisionsToS3RequestDetails) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExportRevisionsToS3RequestDetails) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *ExportRevisionsToS3RequestDetails) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "ExportRevisionsToS3RequestDetails"}
+	if s.DataSetId == nil {
+		invalidParams.Add(request.NewErrParamRequired("DataSetId"))
+	}
+	if s.RevisionDestinations == nil {
+		invalidParams.Add(request.NewErrParamRequired("RevisionDestinations"))
+	}
+	if s.Encryption != nil {
+		if err := s.Encryption.Validate(); err != nil {
+			invalidParams.AddNested("Encryption", err.(request.ErrInvalidParams))
+		}
+	}
+	if s.RevisionDestinations != nil {
+		for i, v := range s.RevisionDestinations {
+			if v == nil {
+				continue
+			}
+			if err := v.Validate(); err != nil {
+				invalidParams.AddNested(fmt.Sprintf("%s[%v]", "RevisionDestinations", i), err.(request.ErrInvalidParams))
+			}
+		}
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetDataSetId sets the DataSetId field's value.
+func (s *ExportRevisionsToS3RequestDetails) SetDataSetId(v string) *ExportRevisionsToS3RequestDetails {
+	s.DataSetId = &v
+	return s
+}
+
+// SetEncryption sets the Encryption field's value.
+func (s *ExportRevisionsToS3RequestDetails) SetEncryption(v *ExportServerSideEncryption) *ExportRevisionsToS3RequestDetails {
+	s.Encryption = v
+	return s
+}
+
+// SetRevisionDestinations sets the RevisionDestinations field's value.
+func (s *ExportRevisionsToS3RequestDetails) SetRevisionDestinations(v []*RevisionDestinationEntry) *ExportRevisionsToS3RequestDetails {
+	s.RevisionDestinations = v
+	return s
+}
+
+// Details about the export revisions to Amazon S3 response.
+type ExportRevisionsToS3ResponseDetails struct {
+	_ struct{} `type:"structure"`
+
+	// The unique identifier for the data set associated with this export job.
+	//
+	// DataSetId is a required field
+	DataSetId *string `type:"string" required:"true"`
+
+	// Encryption configuration of the export job.
+	Encryption *ExportServerSideEncryption `type:"structure"`
+
+	// The destination in Amazon S3 where the revision is exported.
+	//
+	// RevisionDestinations is a required field
+	RevisionDestinations []*RevisionDestinationEntry `type:"list" required:"true"`
+}
+
+// String returns the string representation
+func (s ExportRevisionsToS3ResponseDetails) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s ExportRevisionsToS3ResponseDetails) GoString() string {
+	return s.String()
+}
+
+// SetDataSetId sets the DataSetId field's value.
+func (s *ExportRevisionsToS3ResponseDetails) SetDataSetId(v string) *ExportRevisionsToS3ResponseDetails {
+	s.DataSetId = &v
+	return s
+}
+
+// SetEncryption sets the Encryption field's value.
+func (s *ExportRevisionsToS3ResponseDetails) SetEncryption(v *ExportServerSideEncryption) *ExportRevisionsToS3ResponseDetails {
+	s.Encryption = v
+	return s
+}
+
+// SetRevisionDestinations sets the RevisionDestinations field's value.
+func (s *ExportRevisionsToS3ResponseDetails) SetRevisionDestinations(v []*RevisionDestinationEntry) *ExportRevisionsToS3ResponseDetails {
+	s.RevisionDestinations = v
 	return s
 }
 
@@ -5463,6 +5586,9 @@ type RequestDetails struct {
 	// Details about the export to Amazon S3 request.
 	ExportAssetsToS3 *ExportAssetsToS3RequestDetails `type:"structure"`
 
+	// Details about the export to Amazon S3 request.
+	ExportRevisionsToS3 *ExportRevisionsToS3RequestDetails `type:"structure"`
+
 	// Details about the import from signed URL request.
 	ImportAssetFromSignedUrl *ImportAssetFromSignedUrlRequestDetails `type:"structure"`
 
@@ -5493,6 +5619,11 @@ func (s *RequestDetails) Validate() error {
 			invalidParams.AddNested("ExportAssetsToS3", err.(request.ErrInvalidParams))
 		}
 	}
+	if s.ExportRevisionsToS3 != nil {
+		if err := s.ExportRevisionsToS3.Validate(); err != nil {
+			invalidParams.AddNested("ExportRevisionsToS3", err.(request.ErrInvalidParams))
+		}
+	}
 	if s.ImportAssetFromSignedUrl != nil {
 		if err := s.ImportAssetFromSignedUrl.Validate(); err != nil {
 			invalidParams.AddNested("ImportAssetFromSignedUrl", err.(request.ErrInvalidParams))
@@ -5519,6 +5650,12 @@ func (s *RequestDetails) SetExportAssetToSignedUrl(v *ExportAssetToSignedUrlRequ
 // SetExportAssetsToS3 sets the ExportAssetsToS3 field's value.
 func (s *RequestDetails) SetExportAssetsToS3(v *ExportAssetsToS3RequestDetails) *RequestDetails {
 	s.ExportAssetsToS3 = v
+	return s
+}
+
+// SetExportRevisionsToS3 sets the ExportRevisionsToS3 field's value.
+func (s *RequestDetails) SetExportRevisionsToS3(v *ExportRevisionsToS3RequestDetails) *RequestDetails {
+	s.ExportRevisionsToS3 = v
 	return s
 }
 
@@ -5607,6 +5744,9 @@ type ResponseDetails struct {
 	// Details for the export to Amazon S3 response.
 	ExportAssetsToS3 *ExportAssetsToS3ResponseDetails `type:"structure"`
 
+	// Details for the export revisions to Amazon S3 response.
+	ExportRevisionsToS3 *ExportRevisionsToS3ResponseDetails `type:"structure"`
+
 	// Details for the import from signed URL response.
 	ImportAssetFromSignedUrl *ImportAssetFromSignedUrlResponseDetails `type:"structure"`
 
@@ -5636,6 +5776,12 @@ func (s *ResponseDetails) SetExportAssetsToS3(v *ExportAssetsToS3ResponseDetails
 	return s
 }
 
+// SetExportRevisionsToS3 sets the ExportRevisionsToS3 field's value.
+func (s *ResponseDetails) SetExportRevisionsToS3(v *ExportRevisionsToS3ResponseDetails) *ResponseDetails {
+	s.ExportRevisionsToS3 = v
+	return s
+}
+
 // SetImportAssetFromSignedUrl sets the ImportAssetFromSignedUrl field's value.
 func (s *ResponseDetails) SetImportAssetFromSignedUrl(v *ImportAssetFromSignedUrlResponseDetails) *ResponseDetails {
 	s.ImportAssetFromSignedUrl = v
@@ -5645,6 +5791,70 @@ func (s *ResponseDetails) SetImportAssetFromSignedUrl(v *ImportAssetFromSignedUr
 // SetImportAssetsFromS3 sets the ImportAssetsFromS3 field's value.
 func (s *ResponseDetails) SetImportAssetsFromS3(v *ImportAssetsFromS3ResponseDetails) *ResponseDetails {
 	s.ImportAssetsFromS3 = v
+	return s
+}
+
+// The destination where the assets in the revision will be exported.
+type RevisionDestinationEntry struct {
+	_ struct{} `type:"structure"`
+
+	// The S3 bucket that is the destination for the assets in the revision.
+	//
+	// Bucket is a required field
+	Bucket *string `type:"string" required:"true"`
+
+	// A string representing the pattern for generated names of the individual assets
+	// in the revision. For more information about key patterns, see Key patterns
+	// when exporting revisions (https://docs.aws.amazon.com/data-exchange/latest/userguide/jobs.html#revision-export-keypatterns).
+	KeyPattern *string `type:"string"`
+
+	// The unique identifier for the revision.
+	//
+	// RevisionId is a required field
+	RevisionId *string `type:"string" required:"true"`
+}
+
+// String returns the string representation
+func (s RevisionDestinationEntry) String() string {
+	return awsutil.Prettify(s)
+}
+
+// GoString returns the string representation
+func (s RevisionDestinationEntry) GoString() string {
+	return s.String()
+}
+
+// Validate inspects the fields of the type to determine if they are valid.
+func (s *RevisionDestinationEntry) Validate() error {
+	invalidParams := request.ErrInvalidParams{Context: "RevisionDestinationEntry"}
+	if s.Bucket == nil {
+		invalidParams.Add(request.NewErrParamRequired("Bucket"))
+	}
+	if s.RevisionId == nil {
+		invalidParams.Add(request.NewErrParamRequired("RevisionId"))
+	}
+
+	if invalidParams.Len() > 0 {
+		return invalidParams
+	}
+	return nil
+}
+
+// SetBucket sets the Bucket field's value.
+func (s *RevisionDestinationEntry) SetBucket(v string) *RevisionDestinationEntry {
+	s.Bucket = &v
+	return s
+}
+
+// SetKeyPattern sets the KeyPattern field's value.
+func (s *RevisionDestinationEntry) SetKeyPattern(v string) *RevisionDestinationEntry {
+	s.KeyPattern = &v
+	return s
+}
+
+// SetRevisionId sets the RevisionId field's value.
+func (s *RevisionDestinationEntry) SetRevisionId(v string) *RevisionDestinationEntry {
+	s.RevisionId = &v
 	return s
 }
 
@@ -6899,6 +7109,9 @@ const (
 
 	// TypeExportAssetToSignedUrl is a Type enum value
 	TypeExportAssetToSignedUrl = "EXPORT_ASSET_TO_SIGNED_URL"
+
+	// TypeExportRevisionsToS3 is a Type enum value
+	TypeExportRevisionsToS3 = "EXPORT_REVISIONS_TO_S3"
 )
 
 // Type_Values returns all elements of the Type enum
@@ -6908,5 +7121,6 @@ func Type_Values() []string {
 		TypeImportAssetFromSignedUrl,
 		TypeExportAssetsToS3,
 		TypeExportAssetToSignedUrl,
+		TypeExportRevisionsToS3,
 	}
 }